Re-work `open_resource` and how it handles compression & archives
This commit makes multiple changes to koza.io.utils.open_resource

- Adds support for opening tar files.

- Handles archives (zip and tar) in the same way that the old
`file_archive` source configuration did: it assumes all files in an
archive are of the same format (CSV, JSONL, etc.). It will likely be
future work to allow a way to specify that only certain files in an
archive should be handled.

- Adds more robust checking for gzip compression than checking for a
`.gz` extension.

- open_resource() now returns one or more SizedResource objects that
indicate the size of the resource being opened, and a `.tell()` method
that indicates the position being read in that resource. This will be
necessary to add some sort of progress bar in the future.

- Resources downloaded from the Web now use the same logic as local
files to check for compression/archives.

- Importantly, the resources returned by `open_resource` *are not
automatically closed*. This was inconsistent in the previous version. It
is up to the consumer of the function to explicitly close resources.

- Adds more tests for compressed and archival formats.

- Small typing changes for other koza.io.utils functions, adding
Optional where appropriate