Generalize TransformConfig in order to remove MapTransformConfig
I realized at some point that creating a map from a reader file is just
a type of transform. This change in the configuration makes achieving
that possible.

A map transform is just a transform that relies on two additional
configuration keys: `key` and `values`. To make passing those values in
a YAML config possible, this commit makes it so that any extra fields in
the configuration are parsed into an `extra_fields` field in a
transform.