Remove unused transform `mode` field, add ability to load a module
This allows a transform to be defined as a module (resolvable from
PATH), e.g. `mypackage.transforms.example_transform`, rather than having
to defined it as a file (`/home/user/code/mypackage/transforms/example_transform.py`)

This allows the possibility of creating generic transforms that can be
packaged, installed, and re-used, without having to track down the
filename of the python file where the transform code is located.