Change the API for writing transforms to be decorator-based
In the most recent change to the Koza API implemented in 535bb8a made it
so that transform functions were defined via one of two functions in a
transform script:

1. A serial transform run on every record (called `transform_record`)
2. A single transform run only once (called `transform`)

This changes that API so that transforms do not have to have one of
those specific names-- instead, one function in the transform script
must be decorated by either `@koza.transform()` or
`@koza.transform_record()`.

This also adds two other decorators:

1. `@koza.on_data_begin`, which calls functions just before data starts
   being read.

2. `@koza.on_data_end`, which calls functions after all data has been
   read.

It also adds one other attribute to the `KozaTransform` class (which is
passed to all decorated functions in a transform script): a `state`
object which is an empty dictionary. Along with the `on_data_*`
decorators, this allows maintaining state across calls to
`transform_record`.